requirements.txt
google-cloud-bigquery
vertexai
pandas
